Refactoring, removing CLI
