+++ src/main/java/me/meloni/SolarLogAPI/BasicGUI/Components/Graph/MonthView.java
@@ -1,4 +1,4 @@
-package me.meloni.SolarLogAPI.Interface.Components.Graph;
+package me.meloni.SolarLogAPI.BasicGUI.Components.Graph;
 
 import javax.swing.*;
 import java.awt.*;
