+++ src/main/java/me/meloni/SolarLogAPI/BasicGUI/Components/SimpleFrame.java
@@ -1,4 +1,4 @@
-package me.meloni.SolarLogAPI.Interface;
+package me.meloni.SolarLogAPI.BasicGUI.Components;
 
 import javax.swing.*;
