+++ src/main/java/me/meloni/SolarLogAPI/BasicGUI/GetChosenFile.java
@@ -1,5 +1,6 @@
-package me.meloni.SolarLogAPI.Interface.FileChoosing;
+package me.meloni.SolarLogAPI.BasicGUI;
 
+import me.meloni.SolarLogAPI.BasicGUI.Components.JFileChooserPreset;
 import me.meloni.SolarLogAPI.FileInteraction.GetFile;
 import me.meloni.SolarLogAPI.FileInteraction.ReadFiles.GetDirectory;
 import me.meloni.SolarLogAPI.FileInteraction.ReadFiles.Validate;
