+++ src/main/java/me/meloni/SolarLogAPI/BasicGUI/GraphCustomizer.java
@@ -1,7 +1,9 @@
-package me.meloni.SolarLogAPI.Interface.BasicUI;
+package me.meloni.SolarLogAPI.BasicGUI;
 
+import me.meloni.SolarLogAPI.BasicGUI.Components.Graph.DayCustomizer;
+import me.meloni.SolarLogAPI.BasicGUI.Components.Graph.MonthCustomizer;
 import me.meloni.SolarLogAPI.SolarMap;
-import me.meloni.SolarLogAPI.Interface.SimpleFrame;
+import me.meloni.SolarLogAPI.BasicGUI.Components.SimpleFrame;
 
 import javax.swing.*;
 import java.util.ArrayList;
@@ -26,11 +28,6 @@ public class GraphCustomizer extends JTabbedPane {
         graphFrame.setLocationRelativeTo(f);
         addTab("Day View",new DayCustomizer(data));
         addTab("Month View",new MonthCustomizer(data));
-    }
-
-    public static void run() {
-        SolarMap solarMap = BasicSolarMapCustomizer.solarMap();
-        new GraphCustomizer(solarMap);
     }
 
     public static void setCmp(JComponent c) {
