+++ src/main/java/me/meloni/SolarLogAPI/Handling/APIInitializer.java
@@ -1,7 +1,5 @@
 package me.meloni.SolarLogAPI.Handling;
 
-import me.meloni.SolarLogAPI.Interface.CLI.CMDInterface;
-
 /**
  * This Class is called when the jar package is executed via runtime.
  * @author ChaosMelone9
@@ -13,6 +11,7 @@ public class APIInitializer {
      * @param args arguments passed by runtime
      */
     public static void main(String[] args) {
-        CMDInterface.run(args);
+        String message = "Oh no, looks like you tried to run this package inside the commandline. \n However this is not supposed to run but be used only as a dependency. \n See https://github.com/ChaosMelone9/SolarLogAPI for more info.";
+        System.out.println(message);
     }
 }
